+++ pkg/config/config.go
@@ -31,11 +31,13 @@ var (
 	featRemoteConfigs    = features.Feature("remote-configs")
 	featIntegrationsNext = features.Feature("integrations-next")
 	featDynamicConfig    = features.Feature("dynamic-config")
+	featExtraMetrics     = features.Feature("extra-scrape-metrics")
 
 	allFeatures = []features.Feature{
 		featRemoteConfigs,
 		featIntegrationsNext,
 		featDynamicConfig,
+		featExtraMetrics,
 	}
 )
 
@@ -391,6 +393,10 @@ func load(fs *flag.FlagSet, args []string, loader loaderFunc) (*Config, error) {
 
 	if err := cfg.Integrations.setVersion(version); err != nil {
 		return nil, fmt.Errorf("error loading config file %s: %w", file, err)
+	}
+
+	if features.Enabled(fs, featExtraMetrics) {
+		cfg.Metrics.Global.ExtraMetrics = true
 	}
 
 	// Finally, apply defaults to config that wasn't specified by file or flag